Moves stage functions to separate module
